Address review comments
Move RepositoryServerConnect to repository.go file
Move constants to const.go
Rename ArtifactPrefix to RepoPathPrefix
Remove blob store from comment and logs
