+++ pkg/kopia/command/repository_test.go
@@ -36,10 +36,10 @@ func (s *RepositoryUtilsSuite) TestRepositoryCreateUtil(c *check.C) {
 				ContentCacheMB:  0,
 				MetadataCacheMB: 0,
 				Username:        "test-username",
-				ArtifactPrefix:  "art/prefix",
+				RepoPathPrefix:  "test-path/prefix",
 			},
 			Checker:       check.NotNil,
-			expectedError: "Failed to generate blob store args: unsupported type for the location",
+			expectedError: "Failed to generate storage args: unsupported type for the location",
 		},
 		{
 			cmdArg: RepositoryCommandArgs{
@@ -60,10 +60,25 @@ func (s *RepositoryUtilsSuite) TestRepositoryCreateUtil(c *check.C) {
 				ContentCacheMB:  0,
 				MetadataCacheMB: 0,
 				Username:        "test-username",
-				ArtifactPrefix:  "art/prefix",
+				RepoPathPrefix:  "test-path/prefix",
 			},
-			Checker:     check.IsNil,
-			expectedCmd: []string{"kopia", "--log-level=error", "--config-file=/tmp/config.file", "--log-dir=/tmp/log.dir", "--password=pass123", "repository", "create", "--no-check-for-updates", "--cache-directory=/tmp/cache.dir", "--content-cache-size-mb=0", "--metadata-cache-size-mb=0", "--override-hostname=test-hostname", "--override-username=test-username", "filesystem", "--path=/mnt/data/test-prefix/art/prefix/"},
+			Checker: check.IsNil,
+			expectedCmd: []string{"kopia",
+				"--log-level=error",
+				"--config-file=/tmp/config.file",
+				"--log-dir=/tmp/log.dir",
+				"--password=pass123",
+				"repository",
+				"create",
+				"--no-check-for-updates",
+				"--cache-directory=/tmp/cache.dir",
+				"--content-cache-size-mb=0",
+				"--metadata-cache-size-mb=0",
+				"--override-hostname=test-hostname",
+				"--override-username=test-username",
+				"filesystem",
+				"--path=/mnt/data/test-prefix/test-path/prefix/",
+			},
 		},
 	} {
 		cmd, err := RepositoryCreateCommand(tc.cmdArg)
@@ -98,10 +113,10 @@ func (s *RepositoryUtilsSuite) TestRepositoryConnectUtil(c *check.C) {
 				ContentCacheMB:  0,
 				MetadataCacheMB: 0,
 				Username:        "test-username",
-				ArtifactPrefix:  "art/prefix",
+				RepoPathPrefix:  "test-path/prefix",
 			},
 			Checker:       check.NotNil,
-			expectedError: "Failed to generate blob store args: unsupported type for the location",
+			expectedError: "Failed to generate storage args: unsupported type for the location",
 		},
 		{
 			cmdArg: RepositoryCommandArgs{
@@ -120,11 +135,25 @@ func (s *RepositoryUtilsSuite) TestRepositoryConnectUtil(c *check.C) {
 				CacheDirectory:  "/tmp/cache.dir",
 				ContentCacheMB:  0,
 				MetadataCacheMB: 0,
-				ArtifactPrefix:  "art/prefix",
+				RepoPathPrefix:  "test-path/prefix",
 				PITFlag:         pit,
 			},
-			Checker:     check.IsNil,
-			expectedCmd: []string{"kopia", "--log-level=error", "--config-file=/tmp/config.file", "--log-dir=/tmp/log.dir", "--password=pass123", "repository", "connect", "--no-check-for-updates", "--cache-directory=/tmp/cache.dir", "--content-cache-size-mb=0", "--metadata-cache-size-mb=0", "filesystem", "--path=/mnt/data/test-prefix/art/prefix/", "--point-in-time=1970-01-01T00:00:00.000Z"},
+			Checker: check.IsNil,
+			expectedCmd: []string{"kopia",
+				"--log-level=error",
+				"--config-file=/tmp/config.file",
+				"--log-dir=/tmp/log.dir",
+				"--password=pass123",
+				"repository",
+				"connect",
+				"--no-check-for-updates",
+				"--cache-directory=/tmp/cache.dir",
+				"--content-cache-size-mb=0",
+				"--metadata-cache-size-mb=0",
+				"filesystem",
+				"--path=/mnt/data/test-prefix/test-path/prefix/",
+				"--point-in-time=1970-01-01T00:00:00.000Z",
+			},
 		},
 	} {
 		cmd, err := RepositoryConnectCommand(tc.cmdArg)
@@ -135,4 +164,37 @@ func (s *RepositoryUtilsSuite) TestRepositoryConnectUtil(c *check.C) {
 			c.Assert(strings.Contains(err.Error(), tc.expectedError), check.Equals, true)
 		}
 	}
+}
+
+func (s *RepositoryUtilsSuite) TestRepositoryConnectServerUtil(c *check.C) {
+	cmd := RepositoryConnectServerCommand(RepositoryServerCommandArgs{
+		UserPassword:    "testpass123",
+		ConfigFilePath:  "/tmp/config.file",
+		LogDirectory:    "/tmp/log.dir",
+		CacheDirectory:  "/tmp/cache.dir",
+		Hostname:        "test-hostname",
+		Username:        "test-username",
+		ServerURL:       "https://127.0.0.1:51515",
+		Fingerprint:     "test-fingerprint",
+		ContentCacheMB:  0,
+		MetadataCacheMB: 0,
+	})
+	c.Assert(cmd, check.DeepEquals, []string{"kopia",
+		"--log-level=error",
+		"--config-file=/tmp/config.file",
+		"--log-dir=/tmp/log.dir",
+		"--password=testpass123",
+		"repository",
+		"connect",
+		"server",
+		"--no-check-for-updates",
+		"--no-grpc",
+		"--cache-directory=/tmp/cache.dir",
+		"--content-cache-size-mb=0",
+		"--metadata-cache-size-mb=0",
+		"--override-hostname=test-hostname",
+		"--override-username=test-username",
+		"--url=https://127.0.0.1:51515",
+		"--server-cert-fingerprint=test-fingerprint",
+	})
 }
