+++ pkg/kopia/command/storage/storage_args_test.go
@@ -33,7 +33,7 @@ func (s *StorageUtilsSuite) TestStorageArgsUtil(c *check.C) {
 					},
 				},
 				AssumeRoleDuration: time.Duration(30 * time.Minute),
-				ArtifactPrefix:     "dir/subdir/",
+				RepoPathPrefix:     "dir/subdir/",
 			},
 			Checker:     check.IsNil,
 			expectedCmd: "s3 --bucket=test-bucket --access-key=<****> --secret-access-key=<****> --prefix=test-prefix/dir/subdir/ --disable-tls-verification --region=test-region",
@@ -46,7 +46,7 @@ func (s *StorageUtilsSuite) TestStorageArgsUtil(c *check.C) {
 						typeKey:   "filestore",
 					},
 				},
-				ArtifactPrefix: "dir/subdir",
+				RepoPathPrefix: "dir/subdir",
 			},
 			Checker:     check.IsNil,
 			expectedCmd: "filesystem --path=/mnt/data/test-prefix/dir/subdir/",
@@ -60,7 +60,7 @@ func (s *StorageUtilsSuite) TestStorageArgsUtil(c *check.C) {
 						typeKey:   "gcs",
 					},
 				},
-				ArtifactPrefix: "dir/subdir",
+				RepoPathPrefix: "dir/subdir",
 			},
 			Checker:     check.IsNil,
 			expectedCmd: "gcs --bucket=test-bucket --credentials-file=/tmp/creds.txt --prefix=test-prefix/dir/subdir/",
@@ -82,7 +82,7 @@ func (s *StorageUtilsSuite) TestStorageArgsUtil(c *check.C) {
 						secrets.AzureStorageEnvironment: []byte("AZURECLOUD"),
 					},
 				},
-				ArtifactPrefix: "dir/subdir",
+				RepoPathPrefix: "dir/subdir",
 			},
 			Checker:     check.IsNil,
 			expectedCmd: "azure --container=test-bucket --prefix=test-prefix/dir/subdir/ --storage-account=<****> --storage-key=<****> --storage-domain=blob.core.windows.net",
